More queue updates
Added QueueAdapter, not finished yet. Changes to the current Queue xml
files are from the UI Beta. With QueueAdapter, it also needs another xml
file to finish off the custom list view. Said xml file might need to
change depending on how the Youtube stuff fits in with this.

I'll be working on this super hard after classes on Friday.